pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
psycopg2-binary = "^2.9.10"
mypy = "^1.15.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.4.0"
pyjwt = "^2.10.1"
django-filter = "^25.1"
pytest = "^8.3.5"
requests = "^2.32.3"
factory-boy = "^3.3.3"
pytest-django = "^4.10.0"
pytest-factoryboy = "^2.7.0"
pytest-cov = "^6.0.0"
diff-cover = "^9.2.4"
django-rest-framework-nested = "^0.0.1"
drf-nested-routers = "^0.94.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
